requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
aiobotocore=2.15.2=pypi_0
aiofiles=24.1.0=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aioitertools=0.12.0=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.1=pypi_0
alembic=1.13.3=pypi_0
aniso8601=9.0.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.6.0=pypi_0
appdirs=1.4.4=pypi_0
argcomplete=3.5.1=pypi_0
attrs=24.2.0=pypi_0
bcrypt=4.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blinker=1.8.2=pypi_0
blobconverter=1.4.3=pypi_0
boto3=1.35.38=pypi_0
botocore=1.35.36=pypi_0
brotli=1.1.0=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.8.30=hbcca054_0
cachetools=5.5.0=pypi_0
cffi=1.17.1=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.1.0=pypi_0
coloredlogs=15.0.1=pypi_0
colorlog=6.8.2=pypi_0
common-utils=2.0.1.dev1=pypi_0
contourpy=1.3.0=pypi_0
cryptography=43.0.3=pypi_0
cycler=0.12.1=pypi_0
dacite=1.7.0=pypi_0
databricks-sdk=0.34.0=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.9=pypi_0
dnspython=2.7.0=pypi_0
docker=7.1.0=pypi_0
fiftyone=1.0.0=pypi_0
fiftyone-brain=0.17.0=pypi_0
fiftyone-db=1.1.6=pypi_0
filelock=3.16.1=pypi_0
flask=3.0.3=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.54.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.9.0=pypi_0
ftfy=6.3.0=pypi_0
future=1.0.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
glob2=0.7=pypi_0
google-auth=2.35.0=pypi_0
graphene=3.3=pypi_0
graphql-core=3.2.4=pypi_0
graphql-relay=3.2.0=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.66.2=pypi_0
gunicorn=23.0.0=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
httpcore=1.0.6=pypi_0
httpx=0.27.2=pypi_0
humanfriendly=10.0=pypi_0
humanize=4.11.0=pypi_0
hypercorn=0.17.3=pypi_0
hyperframe=6.0.1=pypi_0
ijson=3.3.0=pypi_0
imageio=2.35.1=pypi_0
importlib-metadata=8.4.0=pypi_0
inflate64=1.0.0=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.4.2=pypi_0
jsonlines=4.0.0=pypi_0
jsonschema=3.2.0=pypi_0
kaleido=0.2.1=pypi_0
kiwisolver=1.4.7=pypi_0
label-studio-sdk=1.0.5=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.43=h712a8e2_1
libexpat=2.6.3=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libgomp=14.1.0=h77fa898_1
libnsl=2.0.1=hd590300_0
libsqlite=3.46.1=hadc24fc_0
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
lightning=2.4.0=pypi_0
lightning-utilities=0.11.7=pypi_0
luxonis-ml=0.4.0=pypi_0
luxonis-train=0.1.0=pypi_0
lxml=5.3.0=pypi_0
mako=1.3.5=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.1=pypi_0
matplotlib=3.9.2=pypi_0
mdurl=0.1.2=pypi_0
mlflow=2.17.0=pypi_0
mlflow-skinny=2.17.0=pypi_0
mongoengine=0.24.2=pypi_0
motor=3.5.3=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multivolumefile=0.2.3=pypi_0
ncurses=6.5=he02047a_1
networkx=3.4=pypi_0
nltk=3.9.1=pypi_0
numpy=2.1.2=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.77=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
onnx=1.17.0=pypi_0
onnxconverter-common=1.14.0=pypi_0
onnxoptimizer=0.3.13=pypi_0
onnxruntime=1.19.2=pypi_0
onnxsim=0.4.36=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openssl=3.3.2=hb9d3cd8_0
opentelemetry-api=1.27.0=pypi_0
opentelemetry-sdk=1.27.0=pypi_0
opentelemetry-semantic-conventions=0.48b0=pypi_0
optuna=4.0.0=pypi_0
optuna-integration=4.0.0=pypi_0
ordered-set=4.1.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
parameterized=0.9.0=pypi_0
paramiko=3.5.0=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=pyh8b19718_1
plotly=5.24.1=pypi_0
polars=1.9.0=pypi_0
pprintpp=0.4.0=pypi_0
priority=2.0.0=pypi_0
propcache=0.2.0=pypi_0
protobuf=3.20.2=pypi_0
psutil=6.0.0=pypi_0
psycopg2-binary=2.9.9=pypi_0
py7zr=0.22.0=pypi_0
pyarrow=17.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pybcj=1.0.2=pypi_0
pycocotools=2.0.8=pypi_0
pycparser=2.22=pypi_0
pycryptodomex=3.21.0=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydantic-settings=2.5.2=pypi_0
pydash=8.0.3=pypi_0
pygments=2.18.0=pypi_0
pymongo=4.8.0=pypi_0
pynacl=1.5.0=pypi_0
pyparsing=3.1.4=pypi_0
pyppmd=1.1.0=pypi_0
pyrsistent=0.20.0=pypi_0
python=3.12.7=hc5c86c4_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytorch-lightning=2.4.0=pypi_0
pytz=2024.2=pypi_0
pyzstd=0.16.2=pypi_0
qudida=0.0.4=pypi_0
rarfile=4.2=pypi_0
readline=8.2=h8228510_1
regex=2024.9.11=pypi_0
requests-mock=1.12.1=pypi_0
retrying=1.3.4=pypi_0
rich=13.9.2=pypi_0
rsa=4.9=pypi_0
s3fs=2024.9.0=pypi_0
s3transfer=0.10.3=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=pyhd8ed1ab_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.35=pypi_0
sqlparse=0.5.1=pypi_0
sse-starlette=0.10.3=pypi_0
sseclient-py=1.8.0=pypi_0
starlette=0.39.2=pypi_0
strawberry-graphql=0.246.1=pypi_0
sympy=1.13.3=pypi_0
tabulate=0.9.0=pypi_0
tenacity=9.0.0=pypi_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
texttable=1.7.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.9.20=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch=2.4.1=pypi_0
torchmetrics=1.4.3=pypi_0
torchvision=0.19.1=pypi_0
tqdm=4.66.5=pypi_0
triton=3.0.0=pypi_0
typeguard=4.3.0=pypi_0
typer=0.12.5=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
tzlocal=5.2=pypi_0
ujson=5.10.0=pypi_0
unique-names-generator=1.0.2=pypi_0
universal-analytics-python3=1.1.1=pypi_0
voxel51-eta=0.13.0=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.4=pypi_0
wheel=0.44.0=pyhd8ed1ab_0
wrapt=1.16.0=pypi_0
wsproto=1.2.0=pypi_0
xmljson=0.2.1=pypi_0
xmltodict=0.14.1=pypi_0
xz=5.2.6=h166bdaf_0
yarl=1.15.2=pypi_0
zipp=3.20.2=pypi_0
